Add lazy initialization in MariaDBStore
- Introduced `lazy_init` parameter to defer initialization of the vector store.
- Added unit tests for lazy initialization scenarios.

Motivation:
The primary motivation for this patch is that,
in a client-server application (e.g., RAG), the developer might not or isn't required to know the embedding_length in advance.
Without lazy init and no prior knowledge of the embedding_length, the developer would be stuck with the default of 1536,
which might differ from what the model actual value is.

Of course, from_texts() will determine the dimension at runtime but we need to consider that
in some cases one would like to get an instance of MariaDBStore first and then perform any other operations later.

Thus, we now support lazy initialization of the store, which will establish the embedding dimension
and create the tables on either the first search or the first insert, whichever comes first.